+++ core/src/test/java/org/elasticsearch/bootstrap/BootstrapCheckTests.java
@@ -63,18 +63,18 @@ public class BootstrapCheckTests extends ESTestCase {
         BoundTransportAddress boundTransportAddress = mock(BoundTransportAddress.class);
         when(boundTransportAddress.boundAddresses()).thenReturn(transportAddresses.toArray(new TransportAddress[0]));
         when(boundTransportAddress.publishAddress()).thenReturn(publishAddress);
-        BootstrapCheck.check(Settings.EMPTY, boundTransportAddress);
+        BootstrapChecks.check(Settings.EMPTY, boundTransportAddress, Collections.emptyList());
     }
 
     public void testNoLogMessageInNonProductionMode() throws NodeValidationException {
         final Logger logger = mock(Logger.class);
-        BootstrapCheck.check(false, Collections.emptyList(), logger);
+        BootstrapChecks.check(false, Collections.emptyList(), logger);
         verifyNoMoreInteractions(logger);
     }
 
     public void testLogMessageInProductionMode() throws NodeValidationException {
         final Logger logger = mock(Logger.class);
-        BootstrapCheck.check(true, Collections.emptyList(), logger);
+        BootstrapChecks.check(true, Collections.emptyList(), logger);
         verify(logger).info("bound or publishing to a non-loopback or non-link-local address, enforcing bootstrap checks");
         verifyNoMoreInteractions(logger);
     }
@@ -98,7 +98,7 @@ public class BootstrapCheckTests extends ESTestCase {
         when(boundTransportAddress.boundAddresses()).thenReturn(transportAddresses.toArray(new TransportAddress[0]));
         when(boundTransportAddress.publishAddress()).thenReturn(publishAddress);
 
-        assertTrue(BootstrapCheck.enforceLimits(boundTransportAddress));
+        assertTrue(BootstrapChecks.enforceLimits(boundTransportAddress));
     }
 
     public void testEnforceLimitsWhenPublishingToNonLocalAddress() {
@@ -114,12 +114,12 @@ public class BootstrapCheckTests extends ESTestCase {
         when(boundTransportAddress.boundAddresses()).thenReturn(transportAddresses.toArray(new TransportAddress[0]));
         when(boundTransportAddress.publishAddress()).thenReturn(publishAddress);
 
-        assertTrue(BootstrapCheck.enforceLimits(boundTransportAddress));
+        assertTrue(BootstrapChecks.enforceLimits(boundTransportAddress));
     }
 
     public void testExceptionAggregation() {
-        final List<BootstrapCheck.Check> checks = Arrays.asList(
-                new BootstrapCheck.Check() {
+        final List<BootstrapCheck> checks = Arrays.asList(
+                new BootstrapCheck() {
                     @Override
                     public boolean check() {
                         return true;
@@ -130,7 +130,7 @@ public class BootstrapCheckTests extends ESTestCase {
                         return "first";
                     }
                 },
-                new BootstrapCheck.Check() {
+                new BootstrapCheck() {
                     @Override
                     public boolean check() {
                         return true;
@@ -144,7 +144,7 @@ public class BootstrapCheckTests extends ESTestCase {
         );
 
         final NodeValidationException e =
-                expectThrows(NodeValidationException.class, () -> BootstrapCheck.check(true, checks, "testExceptionAggregation"));
+                expectThrows(NodeValidationException.class, () -> BootstrapChecks.check(true, checks, "testExceptionAggregation"));
         assertThat(e, hasToString(allOf(containsString("bootstrap checks failed"), containsString("first"), containsString("second"))));
         final Throwable[] suppressed = e.getSuppressed();
         assertThat(suppressed.length, equalTo(2));
@@ -160,7 +160,7 @@ public class BootstrapCheckTests extends ESTestCase {
         final AtomicLong initialHeapSize = new AtomicLong(initial);
         final AtomicLong maxHeapSize = new AtomicLong(max);
 
-        final BootstrapCheck.HeapSizeCheck check = new BootstrapCheck.HeapSizeCheck() {
+        final BootstrapChecks.HeapSizeCheck check = new BootstrapChecks.HeapSizeCheck() {
             @Override
             long getInitialHeapSize() {
                 return initialHeapSize.get();
@@ -175,7 +175,7 @@ public class BootstrapCheckTests extends ESTestCase {
         final NodeValidationException e =
                 expectThrows(
                         NodeValidationException.class,
-                        () -> BootstrapCheck.check(true, Collections.singletonList(check), "testHeapSizeCheck"));
+                        () -> BootstrapChecks.check(true, Collections.singletonList(check), "testHeapSizeCheck"));
         assertThat(
                 e.getMessage(),
                 containsString("initial heap size [" + initialHeapSize.get() + "] " +
@@ -183,7 +183,7 @@ public class BootstrapCheckTests extends ESTestCase {
 
         initialHeapSize.set(maxHeapSize.get());
 
-        BootstrapCheck.check(true, Collections.singletonList(check), "testHeapSizeCheck");
+        BootstrapChecks.check(true, Collections.singletonList(check), "testHeapSizeCheck");
 
         // nothing should happen if the initial heap size or the max
         // heap size is not available
@@ -192,23 +192,23 @@ public class BootstrapCheckTests extends ESTestCase {
         } else {
             maxHeapSize.set(0);
         }
-        BootstrapCheck.check(true, Collections.singletonList(check), "testHeapSizeCheck");
+        BootstrapChecks.check(true, Collections.singletonList(check), "testHeapSizeCheck");
     }
 
     public void testFileDescriptorLimits() throws NodeValidationException {
         final boolean osX = randomBoolean(); // simulates OS X versus non-OS X
         final int limit = osX ? 10240 : 1 << 16;
         final AtomicLong maxFileDescriptorCount = new AtomicLong(randomIntBetween(1, limit - 1));
-        final BootstrapCheck.FileDescriptorCheck check;
+        final BootstrapChecks.FileDescriptorCheck check;
         if (osX) {
-            check = new BootstrapCheck.OsXFileDescriptorCheck() {
+            check = new BootstrapChecks.OsXFileDescriptorCheck() {
                 @Override
                 long getMaxFileDescriptorCount() {
                     return maxFileDescriptorCount.get();
                 }
             };
         } else {
-            check = new BootstrapCheck.FileDescriptorCheck() {
+            check = new BootstrapChecks.FileDescriptorCheck() {
                 @Override
                 long getMaxFileDescriptorCount() {
                     return maxFileDescriptorCount.get();
@@ -218,24 +218,24 @@ public class BootstrapCheckTests extends ESTestCase {
 
         final NodeValidationException e =
                 expectThrows(NodeValidationException.class,
-                        () -> BootstrapCheck.check(true, Collections.singletonList(check), "testFileDescriptorLimits"));
+                        () -> BootstrapChecks.check(true, Collections.singletonList(check), "testFileDescriptorLimits"));
         assertThat(e.getMessage(), containsString("max file descriptors"));
 
         maxFileDescriptorCount.set(randomIntBetween(limit + 1, Integer.MAX_VALUE));
 
-        BootstrapCheck.check(true, Collections.singletonList(check), "testFileDescriptorLimits");
+        BootstrapChecks.check(true, Collections.singletonList(check), "testFileDescriptorLimits");
 
         // nothing should happen if current file descriptor count is
         // not available
         maxFileDescriptorCount.set(-1);
-        BootstrapCheck.check(true, Collections.singletonList(check), "testFileDescriptorLimits");
+        BootstrapChecks.check(true, Collections.singletonList(check), "testFileDescriptorLimits");
     }
 
     public void testFileDescriptorLimitsThrowsOnInvalidLimit() {
         final IllegalArgumentException e =
             expectThrows(
                 IllegalArgumentException.class,
-                () -> new BootstrapCheck.FileDescriptorCheck(-randomIntBetween(0, Integer.MAX_VALUE)));
+                () -> new BootstrapChecks.FileDescriptorCheck(-randomIntBetween(0, Integer.MAX_VALUE)));
         assertThat(e.getMessage(), containsString("limit must be positive but was"));
     }
 
@@ -261,7 +261,7 @@ public class BootstrapCheckTests extends ESTestCase {
         testCases.add(new MlockallCheckTestCase(false, false, false));
 
         for (final MlockallCheckTestCase testCase : testCases) {
-            final BootstrapCheck.MlockallCheck check = new BootstrapCheck.MlockallCheck(testCase.mlockallSet) {
+            final BootstrapChecks.MlockallCheck check = new BootstrapChecks.MlockallCheck(testCase.mlockallSet) {
                 @Override
                 boolean isMemoryLocked() {
                     return testCase.isMemoryLocked;
@@ -271,7 +271,7 @@ public class BootstrapCheckTests extends ESTestCase {
             if (testCase.shouldFail) {
                 final NodeValidationException e = expectThrows(
                         NodeValidationException.class,
-                        () -> BootstrapCheck.check(
+                        () -> BootstrapChecks.check(
                                 true,
                                 Collections.singletonList(check),
                                 "testFileDescriptorLimitsThrowsOnInvalidLimit"));
@@ -280,7 +280,7 @@ public class BootstrapCheckTests extends ESTestCase {
                         containsString("memory locking requested for elasticsearch process but memory is not locked"));
             } else {
                 // nothing should happen
-                BootstrapCheck.check(true, Collections.singletonList(check), "testFileDescriptorLimitsThrowsOnInvalidLimit");
+                BootstrapChecks.check(true, Collections.singletonList(check), "testFileDescriptorLimitsThrowsOnInvalidLimit");
             }
         }
     }
@@ -288,7 +288,7 @@ public class BootstrapCheckTests extends ESTestCase {
     public void testMaxNumberOfThreadsCheck() throws NodeValidationException {
         final int limit = 1 << 11;
         final AtomicLong maxNumberOfThreads = new AtomicLong(randomIntBetween(1, limit - 1));
-        final BootstrapCheck.MaxNumberOfThreadsCheck check = new BootstrapCheck.MaxNumberOfThreadsCheck() {
+        final BootstrapChecks.MaxNumberOfThreadsCheck check = new BootstrapChecks.MaxNumberOfThreadsCheck() {
             @Override
             long getMaxNumberOfThreads() {
                 return maxNumberOfThreads.get();
@@ -297,23 +297,23 @@ public class BootstrapCheckTests extends ESTestCase {
 
         final NodeValidationException e = expectThrows(
                 NodeValidationException.class,
-                () -> BootstrapCheck.check(true, Collections.singletonList(check), "testMaxNumberOfThreadsCheck"));
+                () -> BootstrapChecks.check(true, Collections.singletonList(check), "testMaxNumberOfThreadsCheck"));
         assertThat(e.getMessage(), containsString("max number of threads"));
 
         maxNumberOfThreads.set(randomIntBetween(limit + 1, Integer.MAX_VALUE));
 
-        BootstrapCheck.check(true, Collections.singletonList(check), "testMaxNumberOfThreadsCheck");
+        BootstrapChecks.check(true, Collections.singletonList(check), "testMaxNumberOfThreadsCheck");
 
         // nothing should happen if current max number of threads is
         // not available
         maxNumberOfThreads.set(-1);
-        BootstrapCheck.check(true, Collections.singletonList(check), "testMaxNumberOfThreadsCheck");
+        BootstrapChecks.check(true, Collections.singletonList(check), "testMaxNumberOfThreadsCheck");
     }
 
     public void testMaxSizeVirtualMemory() throws NodeValidationException {
         final long rlimInfinity = Constants.MAC_OS_X ? 9223372036854775807L : -1L;
         final AtomicLong maxSizeVirtualMemory = new AtomicLong(randomIntBetween(0, Integer.MAX_VALUE));
-        final BootstrapCheck.MaxSizeVirtualMemoryCheck check = new BootstrapCheck.MaxSizeVirtualMemoryCheck() {
+        final BootstrapChecks.MaxSizeVirtualMemoryCheck check = new BootstrapChecks.MaxSizeVirtualMemoryCheck() {
             @Override
             long getMaxSizeVirtualMemory() {
                 return maxSizeVirtualMemory.get();
@@ -328,23 +328,23 @@ public class BootstrapCheckTests extends ESTestCase {
 
         final NodeValidationException e = expectThrows(
                 NodeValidationException.class,
-                () -> BootstrapCheck.check(true, Collections.singletonList(check), "testMaxSizeVirtualMemory"));
+                () -> BootstrapChecks.check(true, Collections.singletonList(check), "testMaxSizeVirtualMemory"));
         assertThat(e.getMessage(), containsString("max size virtual memory"));
 
         maxSizeVirtualMemory.set(rlimInfinity);
 
-        BootstrapCheck.check(true, Collections.singletonList(check), "testMaxSizeVirtualMemory");
+        BootstrapChecks.check(true, Collections.singletonList(check), "testMaxSizeVirtualMemory");
 
         // nothing should happen if max size virtual memory is not
         // available
         maxSizeVirtualMemory.set(Long.MIN_VALUE);
-        BootstrapCheck.check(true, Collections.singletonList(check), "testMaxSizeVirtualMemory");
+        BootstrapChecks.check(true, Collections.singletonList(check), "testMaxSizeVirtualMemory");
     }
 
     public void testMaxMapCountCheck() throws NodeValidationException {
         final int limit = 1 << 18;
         final AtomicLong maxMapCount = new AtomicLong(randomIntBetween(1, limit - 1));
-        final BootstrapCheck.MaxMapCountCheck check = new BootstrapCheck.MaxMapCountCheck() {
+        final BootstrapChecks.MaxMapCountCheck check = new BootstrapChecks.MaxMapCountCheck() {
             @Override
             long getMaxMapCount() {
                 return maxMapCount.get();
@@ -353,22 +353,22 @@ public class BootstrapCheckTests extends ESTestCase {
 
         final NodeValidationException e = expectThrows(
                 NodeValidationException.class,
-                () -> BootstrapCheck.check(true, Collections.singletonList(check), "testMaxMapCountCheck"));
+                () -> BootstrapChecks.check(true, Collections.singletonList(check), "testMaxMapCountCheck"));
         assertThat(e.getMessage(), containsString("max virtual memory areas vm.max_map_count"));
 
         maxMapCount.set(randomIntBetween(limit + 1, Integer.MAX_VALUE));
 
-        BootstrapCheck.check(true, Collections.singletonList(check), "testMaxMapCountCheck");
+        BootstrapChecks.check(true, Collections.singletonList(check), "testMaxMapCountCheck");
 
         // nothing should happen if current vm.max_map_count is not
         // available
         maxMapCount.set(-1);
-        BootstrapCheck.check(true, Collections.singletonList(check), "testMaxMapCountCheck");
+        BootstrapChecks.check(true, Collections.singletonList(check), "testMaxMapCountCheck");
     }
 
     public void testClientJvmCheck() throws NodeValidationException {
         final AtomicReference<String> vmName = new AtomicReference<>("Java HotSpot(TM) 32-Bit Client VM");
-        final BootstrapCheck.Check check = new BootstrapCheck.ClientJvmCheck() {
+        final BootstrapCheck check = new BootstrapChecks.ClientJvmCheck() {
             @Override
             String getVmName() {
                 return vmName.get();
@@ -377,19 +377,19 @@ public class BootstrapCheckTests extends ESTestCase {
 
         final NodeValidationException e = expectThrows(
                 NodeValidationException.class,
-                () -> BootstrapCheck.check(true, Collections.singletonList(check), "testClientJvmCheck"));
+                () -> BootstrapChecks.check(true, Collections.singletonList(check), "testClientJvmCheck"));
         assertThat(
                 e.getMessage(),
                 containsString("JVM is using the client VM [Java HotSpot(TM) 32-Bit Client VM] " +
                         "but should be using a server VM for the best performance"));
 
         vmName.set("Java HotSpot(TM) 32-Bit Server VM");
-        BootstrapCheck.check(true, Collections.singletonList(check), "testClientJvmCheck");
+        BootstrapChecks.check(true, Collections.singletonList(check), "testClientJvmCheck");
     }
 
     public void testUseSerialGCCheck() throws NodeValidationException {
         final AtomicReference<String> useSerialGC = new AtomicReference<>("true");
-        final BootstrapCheck.Check check = new BootstrapCheck.UseSerialGCCheck() {
+        final BootstrapCheck check = new BootstrapChecks.UseSerialGCCheck() {
             @Override
             String getUseSerialGC() {
                 return useSerialGC.get();
@@ -398,19 +398,19 @@ public class BootstrapCheckTests extends ESTestCase {
 
         final NodeValidationException e = expectThrows(
             NodeValidationException.class,
-            () -> BootstrapCheck.check(true, Collections.singletonList(check), "testUseSerialGCCheck"));
+            () -> BootstrapChecks.check(true, Collections.singletonList(check), "testUseSerialGCCheck"));
         assertThat(
             e.getMessage(),
             containsString("JVM is using the serial collector but should not be for the best performance; " + "" +
                 "either it's the default for the VM [" + JvmInfo.jvmInfo().getVmName() +"] or -XX:+UseSerialGC was explicitly specified"));
 
         useSerialGC.set("false");
-        BootstrapCheck.check(true, Collections.singletonList(check), "testUseSerialGCCheck");
+        BootstrapChecks.check(true, Collections.singletonList(check), "testUseSerialGCCheck");
     }
 
     public void testSystemCallFilterCheck() throws NodeValidationException {
         final AtomicBoolean isSecompInstalled = new AtomicBoolean();
-        final BootstrapCheck.SystemCallFilterCheck systemCallFilterEnabledCheck = new BootstrapCheck.SystemCallFilterCheck(true) {
+        final BootstrapChecks.SystemCallFilterCheck systemCallFilterEnabledCheck = new BootstrapChecks.SystemCallFilterCheck(true) {
             @Override
             boolean isSeccompInstalled() {
                 return isSecompInstalled.get();
@@ -419,31 +419,31 @@ public class BootstrapCheckTests extends ESTestCase {
 
         final NodeValidationException e = expectThrows(
             NodeValidationException.class,
-            () -> BootstrapCheck.check(true, Collections.singletonList(systemCallFilterEnabledCheck), "testSystemCallFilterCheck"));
+            () -> BootstrapChecks.check(true, Collections.singletonList(systemCallFilterEnabledCheck), "testSystemCallFilterCheck"));
         assertThat(
             e.getMessage(),
             containsString("system call filters failed to install; " +
                 "check the logs and fix your configuration or disable system call filters at your own risk"));
 
         isSecompInstalled.set(true);
-        BootstrapCheck.check(true, Collections.singletonList(systemCallFilterEnabledCheck), "testSystemCallFilterCheck");
+        BootstrapChecks.check(true, Collections.singletonList(systemCallFilterEnabledCheck), "testSystemCallFilterCheck");
 
-        final BootstrapCheck.SystemCallFilterCheck systemCallFilterNotEnabledCheck = new BootstrapCheck.SystemCallFilterCheck(false) {
+        final BootstrapChecks.SystemCallFilterCheck systemCallFilterNotEnabledCheck = new BootstrapChecks.SystemCallFilterCheck(false) {
             @Override
             boolean isSeccompInstalled() {
                 return isSecompInstalled.get();
             }
         };
         isSecompInstalled.set(false);
-        BootstrapCheck.check(true, Collections.singletonList(systemCallFilterNotEnabledCheck), "testSystemCallFilterCheck");
+        BootstrapChecks.check(true, Collections.singletonList(systemCallFilterNotEnabledCheck), "testSystemCallFilterCheck");
         isSecompInstalled.set(true);
-        BootstrapCheck.check(true, Collections.singletonList(systemCallFilterNotEnabledCheck), "testSystemCallFilterCheck");
+        BootstrapChecks.check(true, Collections.singletonList(systemCallFilterNotEnabledCheck), "testSystemCallFilterCheck");
     }
 
     public void testMightForkCheck() throws NodeValidationException {
         final AtomicBoolean isSeccompInstalled = new AtomicBoolean();
         final AtomicBoolean mightFork = new AtomicBoolean();
-        final BootstrapCheck.MightForkCheck check = new BootstrapCheck.MightForkCheck() {
+        final BootstrapChecks.MightForkCheck check = new BootstrapChecks.MightForkCheck() {
             @Override
             boolean isSeccompInstalled() {
                 return isSeccompInstalled.get();
@@ -471,7 +471,7 @@ public class BootstrapCheckTests extends ESTestCase {
     public void testOnErrorCheck() throws NodeValidationException {
         final AtomicBoolean isSeccompInstalled = new AtomicBoolean();
         final AtomicReference<String> onError = new AtomicReference<>();
-        final BootstrapCheck.MightForkCheck check = new BootstrapCheck.OnErrorCheck() {
+        final BootstrapChecks.MightForkCheck check = new BootstrapChecks.OnErrorCheck() {
             @Override
             boolean isSeccompInstalled() {
                 return isSeccompInstalled.get();
@@ -499,7 +499,7 @@ public class BootstrapCheckTests extends ESTestCase {
     public void testOnOutOfMemoryErrorCheck() throws NodeValidationException {
         final AtomicBoolean isSeccompInstalled = new AtomicBoolean();
         final AtomicReference<String> onOutOfMemoryError = new AtomicReference<>();
-        final BootstrapCheck.MightForkCheck check = new BootstrapCheck.OnOutOfMemoryErrorCheck() {
+        final BootstrapChecks.MightForkCheck check = new BootstrapChecks.OnOutOfMemoryErrorCheck() {
             @Override
             boolean isSeccompInstalled() {
                 return isSeccompInstalled.get();
@@ -526,7 +526,7 @@ public class BootstrapCheckTests extends ESTestCase {
     }
 
     private void runMightForkTest(
-        final BootstrapCheck.MightForkCheck check,
+        final BootstrapChecks.MightForkCheck check,
         final AtomicBoolean isSeccompInstalled,
         final Runnable disableMightFork,
         final Runnable enableMightFork,
@@ -541,13 +541,13 @@ public class BootstrapCheckTests extends ESTestCase {
         } else {
             enableMightFork.run();
         }
-        BootstrapCheck.check(true, Collections.singletonList(check), methodName);
+        BootstrapChecks.check(true, Collections.singletonList(check), methodName);
 
         // if seccomp is enabled, but we will not fork, nothing should
         // happen
         isSeccompInstalled.set(true);
         disableMightFork.run();
-        BootstrapCheck.check(true, Collections.singletonList(check), methodName);
+        BootstrapChecks.check(true, Collections.singletonList(check), methodName);
 
         // if seccomp is enabled, and we might fork, the check should
         // be enforced, regardless of bootstrap checks being enabled or
@@ -557,7 +557,7 @@ public class BootstrapCheckTests extends ESTestCase {
 
         final NodeValidationException e = expectThrows(
             NodeValidationException.class,
-            () -> BootstrapCheck.check(randomBoolean(), Collections.singletonList(check), methodName));
+            () -> BootstrapChecks.check(randomBoolean(), Collections.singletonList(check), methodName));
         consumer.accept(e);
     }
 
@@ -566,7 +566,7 @@ public class BootstrapCheckTests extends ESTestCase {
         final AtomicBoolean isJava8 = new AtomicBoolean(true);
         final AtomicReference<String> jvmVersion =
             new AtomicReference<>(String.format(Locale.ROOT, "25.%d-b%d", randomIntBetween(0, 39), randomIntBetween(1, 128)));
-        final BootstrapCheck.G1GCCheck oracleCheck = new BootstrapCheck.G1GCCheck() {
+        final BootstrapChecks.G1GCCheck oracleCheck = new BootstrapChecks.G1GCCheck() {
 
             @Override
             String jvmVendor() {
@@ -593,7 +593,7 @@ public class BootstrapCheckTests extends ESTestCase {
         final NodeValidationException e =
             expectThrows(
                 NodeValidationException.class,
-                () -> BootstrapCheck.check(true, Collections.singletonList(oracleCheck), "testG1GCCheck"));
+                () -> BootstrapChecks.check(true, Collections.singletonList(oracleCheck), "testG1GCCheck"));
         assertThat(
             e.getMessage(),
             containsString(
@@ -601,14 +601,14 @@ public class BootstrapCheckTests extends ESTestCase {
 
         // if G1GC is disabled, nothing should happen
         isG1GCEnabled.set(false);
-        BootstrapCheck.check(true, Collections.singletonList(oracleCheck), "testG1GCCheck");
+        BootstrapChecks.check(true, Collections.singletonList(oracleCheck), "testG1GCCheck");
 
         // if on or after update 40, nothing should happen independent of whether or not G1GC is enabled
         isG1GCEnabled.set(randomBoolean());
         jvmVersion.set(String.format(Locale.ROOT, "25.%d-b%d", randomIntBetween(40, 112), randomIntBetween(1, 128)));
-        BootstrapCheck.check(true, Collections.singletonList(oracleCheck), "testG1GCCheck");
+        BootstrapChecks.check(true, Collections.singletonList(oracleCheck), "testG1GCCheck");
 
-        final BootstrapCheck.G1GCCheck nonOracleCheck = new BootstrapCheck.G1GCCheck() {
+        final BootstrapChecks.G1GCCheck nonOracleCheck = new BootstrapChecks.G1GCCheck() {
 
             @Override
             String jvmVendor() {
@@ -618,9 +618,9 @@ public class BootstrapCheckTests extends ESTestCase {
         };
 
         // if not on an Oracle JVM, nothing should happen
-        BootstrapCheck.check(true, Collections.singletonList(nonOracleCheck), "testG1GCCheck");
+        BootstrapChecks.check(true, Collections.singletonList(nonOracleCheck), "testG1GCCheck");
 
-        final BootstrapCheck.G1GCCheck nonJava8Check = new BootstrapCheck.G1GCCheck() {
+        final BootstrapChecks.G1GCCheck nonJava8Check = new BootstrapChecks.G1GCCheck() {
 
             @Override
             boolean isJava8() {
@@ -630,11 +630,11 @@ public class BootstrapCheckTests extends ESTestCase {
         };
 
         // if not Java 8, nothing should happen
-        BootstrapCheck.check(true, Collections.singletonList(nonJava8Check), "testG1GCCheck");
+        BootstrapChecks.check(true, Collections.singletonList(nonJava8Check), "testG1GCCheck");
     }
 
     public void testAlwaysEnforcedChecks() {
-        final BootstrapCheck.Check check = new BootstrapCheck.Check() {
+        final BootstrapCheck check = new BootstrapCheck() {
             @Override
             public boolean check() {
                 return true;
@@ -653,7 +653,7 @@ public class BootstrapCheckTests extends ESTestCase {
 
         final NodeValidationException alwaysEnforced = expectThrows(
             NodeValidationException.class,
-            () -> BootstrapCheck.check(randomBoolean(), Collections.singletonList(check), "testAlwaysEnforcedChecks"));
+            () -> BootstrapChecks.check(randomBoolean(), Collections.singletonList(check), "testAlwaysEnforcedChecks"));
         assertThat(alwaysEnforced, hasToString(containsString("error")));
     }
