+++ test/framework/src/main/java/org/elasticsearch/node/NodeTests.java
@@ -20,17 +20,23 @@ package org.elasticsearch.node;
 
 import org.apache.logging.log4j.Logger;
 import org.elasticsearch.Version;
+import org.elasticsearch.bootstrap.BootstrapCheck;
 import org.elasticsearch.cluster.ClusterName;
 import org.elasticsearch.common.network.NetworkModule;
 import org.elasticsearch.common.settings.Settings;
+import org.elasticsearch.common.transport.BoundTransportAddress;
 import org.elasticsearch.env.Environment;
+import org.elasticsearch.plugins.Plugin;
 import org.elasticsearch.test.ESTestCase;
 import org.elasticsearch.test.InternalTestCluster;
 import org.elasticsearch.transport.MockTcpTransportPlugin;
 
 import java.io.IOException;
 import java.nio.file.Path;
+import java.util.Arrays;
 import java.util.Collections;
+import java.util.List;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 import static org.hamcrest.Matchers.equalTo;
 import static org.mockito.Mockito.mock;
@@ -59,6 +65,52 @@ public class NodeTests extends ESTestCase {
             } else {
                 assertThat(Node.NODE_NAME_SETTING.get(nodeSettings), equalTo(name));
             }
+        }
+    }
+
+    public static class CheckPlugin extends Plugin {
+        public static final BootstrapCheck CHECK = new BootstrapCheck() {
+            @Override
+            public boolean check() {
+                return false;
+            }
+
+            @Override
+            public String errorMessage() {
+                return "boom";
+            }
+        };
+        @Override
+        public List<BootstrapCheck> getBootstrapChecks() {
+            return Collections.singletonList(CHECK);
+        }
+    }
+
+    public void testLoadPluginBootstrapChecks() throws IOException {
+        final Path tempDir = createTempDir();
+        final String name = randomBoolean() ? randomAsciiOfLength(10) : null;
+        Settings.Builder settings = Settings.builder()
+            .put(ClusterName.CLUSTER_NAME_SETTING.getKey(), InternalTestCluster.clusterName("single-node-cluster", randomLong()))
+            .put(Environment.PATH_HOME_SETTING.getKey(), tempDir)
+            .put(NetworkModule.HTTP_ENABLED.getKey(), false)
+            .put("transport.type", "mock-socket-network")
+            .put(Node.NODE_DATA_SETTING.getKey(), true);
+        if (name != null) {
+            settings.put(Node.NODE_NAME_SETTING.getKey(), name);
+        }
+        AtomicBoolean executed = new AtomicBoolean(false);
+        try (Node node = new MockNode(settings.build(), Arrays.asList(MockTcpTransportPlugin.class, CheckPlugin.class)) {
+            @Override
+            protected void validateNodeBeforeAcceptingRequests(Settings settings, BoundTransportAddress boundTransportAddress,
+                                                               List<BootstrapCheck> bootstrapChecks) throws NodeValidationException {
+                assertEquals(1, bootstrapChecks.size());
+                assertSame(CheckPlugin.CHECK, bootstrapChecks.get(0));
+                executed.set(true);
+                throw new NodeValidationException("boom");
+            }
+        }) {
+            expectThrows(NodeValidationException.class, () -> node.start());
+            assertTrue(executed.get());
         }
     }
